add tabs layout
