+++ app/app/(tabs)/bookmarks.tsx
@@ -1,0 +1,5 @@
+import { View } from "react-native-ui-lib";
+
+export default function Bookmarks() {
+  return <View />;
+}
