+++ app/app/(tabs)/settings.tsx
@@ -1,0 +1,5 @@
+import { View } from "react-native-ui-lib";
+
+export default function Settings() {
+  return <View />;
+}
