+++ app/utils/theme.ts
@@ -1,0 +1,1 @@
+export const PRIMARY_COLOR = "#388659";
